requirements.txt
facexlib>=0.2.5
numpy
opencv-python
Pillow
torch>=1.7
torchvision
tqdm
streamlit
opencv-python
new-gfpgan==1.0.1
MY-basicsr==1.4.2
pyrebase4
streamlit-extras
streamlit-image-comparison